architecture du projet et creation des composants
