+++ client/src/App/assets/fonts/Lato-Bold.ttf
