+++ client/src/App/assets/fonts/Lato-BoldItalic.ttf
