+++ client/src/App/assets/fonts/Lato-Italic.ttf
