+++ client/src/App/assets/fonts/Lato-LightItalic.ttf
