+++ client/src/App/assets/fonts/Lato-Regular.ttf
