+++ client/src/App/assets/fonts/icons/config.json
