+++ client/src/App/assets/img/Nidalee.png
