+++ client/src/App/components/button/button.scss
@@ -1,0 +1,26 @@
+.group-btn {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.btn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 104px;
+    height: 32px;
+    color: $blue;
+    font-size: 16px;
+    cursor: pointer;
+    user-select: none;
+    &:not(.active) {
+        border: solid 2px;
+        border-image: linear-gradient(#C79A3C, #A6843C, #614B1D) 2;
+    }
+    &:hover,
+    &.active {
+        background: $gold-gradient;
+        color: $white;
+    }
+}
