+++ client/src/App/components/champbubble/champbubble.scss
@@ -1,0 +1,8 @@
+.thumb-champ {
+    width: 80px;
+    height: 80px;
+    border-radius: 50%;
+    box-shadow: $box-shadow;
+    background-size: cover;
+    background-repeat: no-repeat;
+}
