--- client/src/App/postChamp/Filter.js
@@ -1,19 +1,0 @@
-
-const JsonLength = (array) => {
-    return array.length
-}
-
-const PostChamp = (array) => {
-    let champions = array.map((champion, index) => {
-        return {
-            poste: champion.poste,
-            name: champion.name
-        }
-    })
-    return champions
-}
-
-module.exports = {
-    JsonLength,
-    PostChamp
-}
