--- client/src/App/postChamp/postChamp.jsx
@@ -1,20 +1,0 @@
-import React, { Component } from 'react';
-import Json from '../../data.json'
-import Filter from './Filter'
-
-class postChamp extends Component {
-   
-    componentDidMount() {
-        console.log(Filter.PostChamp(Json))
-    }
-    render() {
-
-        return (
-            <div>
-                
-            </div>
-        )
-    }
-}
-
-export default postChamp
