+++ client/src/App/scripts/SortByRate.spec.js
