--- client/src/App/style.scss
@@ -1,290 +1,0 @@
-$background : radial-gradient(rgba(0, 43, 84, 0.9) 0.97%,rgba(12, 12, 36, 0.9) 56.75%,rgba(15, 15, 46, 0.9) 100%);
-$darkblue: #160B31;
-$blue: #00CBE0;
-$blue-gradient: linear-gradient(180deg, #00CBE0 0%, rgba(0, 203, 224, 0.2) 98.08%);
-$gold: #C79A3C;
-$gold-gradient: linear-gradient(180deg, #C79A3C 0%, #A6843C 39.06%, #614B1D 100%);
-$red-gradient: linear-gradient(180deg, #FC0044 0%, rgba(252, 0, 68, 0.2) 98.08%);
-$white: #ffffff;
-$black: #000000;
-
-$box-shadow: 0px 0px 30px rgba(0, 203, 224, 0.15);
-
-$font: 'Lato';
-$font-bold: 'Lato-bold';
-
-@font-face {
-    font-family: 'fontello';
-    src: url('fonts/icons/fontello.eot?73249035');
-    src: url('fonts/icons/fontello.eot?73249035#iefix') format('embedded-opentype'),
-         url('fonts/icons/fontello.woff2?73249035') format('woff2'),
-         url('fonts/icons/fontello.woff?73249035') format('woff'),
-         url('fonts/icons/fontello.ttf?73249035') format('truetype'),
-         url('fonts/icons/fontello.svg?73249035#fontello') format('svg');
-    font-weight: normal;
-    font-style: normal;
-  }
-
-@font-face {
-    font-family: "lato";
-    src: url(./fonts/Lato-Regular.ttf);
-}
-
-@font-face {
-    font-family: "lato-hairline";
-    src: url(./fonts/Lato-Hairline.ttf);
-}
-
-@font-face {
-    font-family: "lato-light";
-    src: url(./fonts/Lato-Light.ttf);
-}
-
-@font-face {
-    font-family: "lato-bold";
-    src: url(./fonts/Lato-Bold.ttf);
-}
-
-html, body {
-    height: 100%;
-}
-
-body {
-    background: $background;
-    font-family: $font;
-    font-size: 14px;
-    color: $white;
-}
-
-h1 {
-    margin-bottom: 20px;
-}
-
-.title {
-    font-family: $font-bold;
-    font-size: 24px;
-}
-
-.wrapper {
-    max-width: 1040px;
-    margin: 0 auto;
-}
-
-.container {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    margin: 20px 0;
-}
-
-.block {
-    width: 100px;
-    height: 100px;
-    border-radius: 8px;
-    margin-right: 20px;
-    
-    &.darkblue {
-        background: $darkblue;
-    }
-    &.blue {
-        background: $blue;
-    }
-    &.gold {
-        background: $gold;
-    }
-    &.blue-gradient {
-        background: $blue-gradient;
-    }
-    &.gold-gradient {
-        background: $gold-gradient;
-    }
-    &.red-gradient {
-        background: $red-gradient;
-    }
-}
-
-.thumb-champ {
-    width: 80px;
-    height: 80px;
-    border-radius: 50%;
-    box-shadow: $box-shadow;
-    background-size: cover;
-    &:nth-child(1) {
-    background-image: url(./img/Gnar.png);
-    }
-    &:nth-child(2) {
-        background-image: url(./img/Nidalee.png);
-    }
-    &:nth-child(3) {
-        background-image: url(./img/Fizz.png);
-    }
-    &:nth-child(4) {
-        background-image: url(./img/Twitch.png);
-    }
-    &:nth-child(5) {
-        background-image: url(./img/Lulu.png);
-    }
-}
-
-.thumb-post {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: flex-end;
-    background-color: $darkblue;
-    width: 120px;
-    height: 120px;
-    border-radius: 50%;
-    box-shadow: $box-shadow;
-    &.icon:before {
-        top: 30px;
-    }
-    p {
-        font-family: $font-bold;
-        text-transform: uppercase;
-        color: $blue;
-        font-size: 18px;
-        margin-bottom: 30px;
-    }
-}
-
-.navigation {
-    z-index: 10;
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    width: 64px;
-    background: $darkblue;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    padding: 48px 0;
-    .logo {
-        width: 32px;
-        user-select: none;
-        img {
-            width: 100%;
-        }
-    }
-    .btnRank {
-        cursor: pointer;
-        position: relative;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 40px;
-        height: 23.09px;
-        user-select: none;
-        &:not(:last-child) {
-            margin-bottom: 32px;
-        }
-        &:hover {
-            span {
-                transition: ease-in-out 0.3s;
-                color: $gold;
-            }
-        }
-        &.active {
-            border-left: solid 3px $gold;
-            border-right: solid 3px $gold;
-            &:before,
-            &:after {
-                content: "";
-                position: absolute;
-                z-index: 1;
-                width: 28.28px;
-                height: 28.28px;
-                transform: scaleY(0.5774) rotate(-45deg);
-                background-color: inherit;
-                left: 2.8579px;
-            }
-            &:before {
-                top: -14.1421px;
-                border-top: solid 4.2426px $gold;
-                border-right: solid 4.2426px $gold;
-            }
-            &:after {
-                bottom: -14.1421px;
-                border-bottom: solid 4.2426px $gold;
-                border-left: solid 4.2426px $gold;
-            }
-            span {
-                color: $gold;
-            }
-        }
-        span {
-            z-index: 2;
-            font-family: $font-bold;
-            font-size: 20px;
-        }
-    }
-    .soundbox {
-        user-select: none;
-        img {
-            cursor: pointer;
-        }
-    }
-}
-
-.group-btn {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 104px;
-    height: 32px;
-    color: $blue;
-    font-size: 16px;
-    cursor: pointer;
-    user-select: none;
-    &:not(.active) {
-        border: solid 2px;
-        border-image: linear-gradient(#C79A3C, #A6843C, #614B1D) 2;
-    }
-    &.active {
-        background: $gold-gradient;
-        color: $white;
-    }
-}
-
-.icons .icon {
-    width: 40px;
-    height: 40px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.icon {
-    position: relative;
-
-    &:before {
-        position: absolute;
-        font-family: 'fontello';
-        color: $blue;
-        font-size: 32px;
-        user-select: none;
-    }
-    &-top:before {
-        content: '\E801';
-    }
-    &-jgl:before {
-        content: '\E800';
-    }
-    &-mid:before {
-        content: '\E803';
-    }
-    &-bot:before {
-        content: '\E804';
-    }
-    &-supp:before {
-        content: '\E802';
-    }
-}
