+++ client/src/App/styles/_global.scss
@@ -1,0 +1,127 @@
+@font-face {
+    font-family: 'fontello';
+    src: url('../assets/fonts/icons/fontello.eot?73249035');
+    src: url('../assets/fonts/icons/fontello.eot?73249035#iefix') format('embedded-opentype'),
+         url('../assets/fonts/icons/fontello.woff2?73249035') format('woff2'),
+         url('../assets/fonts/icons/fontello.woff?73249035') format('woff'),
+         url('../assets/fonts/icons/fontello.ttf?73249035') format('truetype'),
+         url('../assets/fonts/icons/fontello.svg?73249035#fontello') format('svg');
+    font-weight: normal;
+    font-style: normal;
+  }
+
+@font-face {
+    font-family: "lato";
+    src: url(../assets/fonts/Lato-Regular.ttf);
+}
+
+@font-face {
+    font-family: "lato-hairline";
+    src: url(../assets/fonts/Lato-Hairline.ttf);
+}
+
+@font-face {
+    font-family: "lato-light";
+    src: url(../assets/fonts/Lato-Light.ttf);
+}
+
+@font-face {
+    font-family: "lato-bold";
+    src: url(../assets/fonts/Lato-Bold.ttf);
+}
+
+html, body {
+    height: 100%;
+}
+
+body {
+    background: $background;
+    background-size: cover;
+    background-repeat: no-repeat;
+    font-family: $font;
+    font-size: 14px;
+    color: $white;
+}
+
+h1 {
+    margin-bottom: 20px;
+}
+
+.title {
+    font-family: $font-bold;
+    font-size: 24px;
+}
+
+.wrapper {
+    max-width: 1040px;
+    margin: 0 auto;
+}
+
+.container {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    margin: 20px 0;
+}
+
+.block {
+    width: 100px;
+    height: 100px;
+    border-radius: 8px;
+    margin-right: 20px;
+    
+    &.darkblue {
+        background: $darkblue;
+    }
+    &.blue {
+        background: $blue;
+    }
+    &.gold {
+        background: $gold;
+    }
+    &.blue-gradient {
+        background: $blue-gradient;
+    }
+    &.gold-gradient {
+        background: $gold-gradient;
+    }
+    &.red-gradient {
+        background: $red-gradient;
+    }
+}
+
+.icons .icon {
+    width: 40px;
+    height: 40px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+
+.icon {
+    position: relative;
+
+    &:before {
+        position: absolute;
+        font-family: 'fontello';
+        color: $blue;
+        font-size: 32px;
+        user-select: none;
+    }
+    &-top:before {
+        content: '\E801';
+    }
+    &-jgl:before {
+        content: '\E800';
+    }
+    &-mid:before {
+        content: '\E803';
+    }
+    &-bot:before {
+        content: '\E804';
+    }
+    &-supp:before {
+        content: '\E802';
+    }
+}
