+++ client/src/App/styles/_variables.scss
@@ -1,0 +1,14 @@
+$background : radial-gradient(rgba(0, 43, 84, 0.9) 0.97%,rgba(12, 12, 36, 0.9) 56.75%,rgba(15, 15, 46, 0.9) 100%);
+$darkblue: #160B31;
+$blue: #00CBE0;
+$blue-gradient: linear-gradient(180deg, #00CBE0 0%, rgba(0, 203, 224, 0.2) 98.08%);
+$gold: #C79A3C;
+$gold-gradient: linear-gradient(180deg, #C79A3C 0%, #A6843C 39.06%, #614B1D 100%);
+$red-gradient: linear-gradient(180deg, #FC0044 0%, rgba(252, 0, 68, 0.2) 98.08%);
+$white: #ffffff;
+$black: #000000;
+
+$box-shadow: 0px 0px 30px rgba(0, 203, 224, 0.15);
+
+$font: 'Lato';
+$font-bold: 'Lato-bold';
