--- client/src/Components/Graph/Graph.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-// import SortByRate from './SortByRate'
-// import data from '../../data.json'
-class Graph extends Component { 
-
-    render() {
-        return(
-            <div></div>
-        )
-    }
-}
-
-export default Graph
